requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0
pydantic==2.6.4
httplib2==0.22.0
